chore(connect-common): move VERSIONs constants
Moving VERSIONS constants file in connect to connect-common since those
constants are used in connect, connect-web and connect-mobile.

Also had to chcange the script  connect/scripts/bump-version.ts so the
versions are updated properly.
