+++ packages/connect-common/src/data/index.ts
@@ -1,0 +1,1 @@
+export * from './version';
