+++ packages/connect-common/src/data/version.ts
